add housing community auth

+++ src/main/java/pl/edu/pg/eti/kask/kzawora/user/UserService.java
@@ -1,13 +1,14 @@
 package pl.edu.pg.eti.kask.kzawora.user;
 
-import pl.edu.pg.eti.kask.kzawora.user.model.PersonalData;
 import pl.edu.pg.eti.kask.kzawora.user.model.User;
 
 import javax.ejb.Stateless;
+import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
-import java.util.ArrayList;
+import java.security.AccessControlException;
 import java.util.List;
 
 @Stateless
@@ -15,37 +16,49 @@ public class UserService {
     @PersistenceContext
     private EntityManager em;
 
+    /**
+     * Nasty hack. SecurityContext can not be injected when legacy domain in WildFly is used.
+     * <p>
+     * Using ExternalContext here causes REST and Servlet won't work.
+     * <p>
+     * Using REST SecurityContext causes JSF won't work.
+     * <p>
+     * Using SessionContext requires making this bean EJB.
+     */
+    @Inject
+    private HttpServletRequest securityContext;
+
     public UserService() {
     }
 
-    @Transactional
-    public void init() {
-        /*
-        List<User> users = new ArrayList<>();
-        users.add(new User(new PersonalData("Nowak", "Jan", "1211111111"), "jan1.nowak@nowak.com", "raz1dwatrzy"));
-        users.add(new User(new PersonalData("Kowalski", "Janek", "4111111111"), "jan2.nowak@nowak.com", "razdwa2trzy"));
-        users.add(new User(new PersonalData("Nowakowski", "Janko", "5111111111"), "jan3.nowak@nowak.com", "razdwatrzy3"));
-        users.add(new User(new PersonalData("Kowal", "Jaś", "1111111611"), "jan4.nowak@nowak.com", "raz1dwa2trzy"));
-        users.add(new User(new PersonalData("Nowy", "Jasiek", "1111131111"), "jan5.nowak@nowak.com", "raz1dwa2trzy3"));
-        users.add(new User(new PersonalData("Nowikowski", "Janek", "4111111111"), "jan6.nowak@nowak.com", "razdwa4trzy3"));
-        users.add(new User(new PersonalData("Ryszard", "Janko", "1116111111"), "jan7.nowak@nowak.com", "razdwat4rzy"));
-        users.add(new User(new PersonalData("Kowalewski", "Jaś", "1111111111"), "jan8.nowak@nowak.com", "ra4zd1w2at3rzy"));
-        for (User u : users) {
-            em.persist(u);
-        }
-*/
-    }
 
     public synchronized List<User> findAllUsers() {
-        return em.createNamedQuery(User.Queries.FIND_ALL, User.class).getResultList();
-
+        if (securityContext.isUserInRole(User.Roles.USER)) {
+            if (securityContext.isUserInRole(User.Roles.ADMIN)) {
+                return em.createNamedQuery(User.Queries.FIND_ALL, User.class).getResultList();
+            } else {
+                String callerUsername = securityContext.getUserPrincipal().toString();
+                return List.of(findUser(callerUsername));
+            }
+        } else {
+            throw new AccessControlException("Access denied");
+        }
     }
 
     public synchronized List<User> findAllUsers(int offset, int limit) {
-        return em.createNamedQuery(User.Queries.FIND_ALL, User.class)
-                .setFirstResult(offset)
-                .setMaxResults(limit)
-                .getResultList();
+        if (securityContext.isUserInRole(User.Roles.USER)) {
+            if (securityContext.isUserInRole(User.Roles.ADMIN)) {
+                return em.createNamedQuery(User.Queries.FIND_ALL, User.class)
+                        .setFirstResult(offset)
+                        .setMaxResults(limit)
+                        .getResultList();
+            } else {
+                String callerUsername = securityContext.getUserPrincipal().toString();
+                return List.of(findUser(callerUsername));
+            }
+        } else {
+            throw new AccessControlException("Access denied");
+        }
     }
 
     public synchronized long countUsers() {
@@ -53,20 +66,61 @@ public class UserService {
     }
 
     public synchronized User findUser(int id) {
-        return em.find(User.class, id);
+        String callerUsername = securityContext.getUserPrincipal().toString();
+        if (securityContext.isUserInRole(User.Roles.USER)) {
+            if (findUser(callerUsername).getId() == id || securityContext.isUserInRole(User.Roles.ADMIN))
+                return em.find(User.class, id);
+            else
+                throw new AccessControlException("Access denied");
+        } else {
+            throw new AccessControlException("Access denied");
+        }
     }
+
+    public synchronized User findUser(String login) {
+        if (securityContext.isUserInRole(User.Roles.USER)) {
+            return em.createNamedQuery(User.Queries.FIND_BY_NAME, User.class).setParameter("login", login).getSingleResult();
+        } else {
+            throw new AccessControlException("Access denied");
+        }
+    }
+
 
     @Transactional
     public synchronized void saveUser(User user) {
-        if (user.getId() == null) {
-            em.persist(user);
+        if (securityContext.isUserInRole(User.Roles.USER)) {
+            String callerUsername = securityContext.getUserPrincipal().toString();
+            if (user.getId() == null) {
+                if (securityContext.isUserInRole(User.Roles.ADMIN)) {
+                    em.persist(user);
+                } else {
+                    throw new AccessControlException("Access denied");
+                }
+            } else {
+                if (callerUsername.equals(user.getLogin())) {
+                    User u = findUser(callerUsername);
+                    if (u.getRoles().containsAll(user.getRoles()))
+                        em.merge(user);
+                    else
+                        throw new AccessControlException("Access denied");
+                } else if (securityContext.isUserInRole(User.Roles.ADMIN)) {
+                    em.merge(user);
+                } else {
+                    throw new AccessControlException("Access denied");
+                }
+            }
         } else {
-            em.merge(user);
+            throw new AccessControlException("Access denied");
         }
     }
 
     @Transactional
     public void removeUser(User user) {
-        em.remove(em.merge(user));
+        String callerUsername = securityContext.getUserPrincipal().toString();
+        if (securityContext.isUserInRole(User.Roles.ADMIN) && !callerUsername.equals(user.getLogin())) {
+            em.remove(em.merge(user));
+        } else {
+            throw new AccessControlException("Access denied");
+        }
     }
 }
